fixes #31673 - Fixing error handling when host is not in build mode
With current implementation, it's possible to download a bootdisk
full host even for hosts that are not in build mode. So this scenario
ends with the Foreman error every time for hosts that are not in build mode.
This PR handles this error much better. It adds some error messages,
refuses serving an image for API requests and also disables button for download.